refactor cnn backbone and make configurable
